backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
semgrep==1.75.0
bandit==1.7.9
reportlab==4.2.0
python-dotenv==1.0.1
pydantic==2.7.1
aiofiles==23.2.1
websockets==12.0
python-magic==0.4.27
requests==2.32.3
sslyze
termcolor==2.4.0
cprint
pycryptodomex==3.20.0
PyYAML>=6.0.1
python-jose[cryptography]
bcrypt==3.2.2
passlib[bcrypt]
python-multipart
google-auth
